requirements.txt
PyYAML==6.0.2
Jinja2==3.1.6
MarkupSafe==3.0.2